Basic ShoppingList (db) functionality
codebar in articleItem to Long.
Improve Dao and Repository funtions.
Demo in DetailArticle

+++ app/src/main/java/com/alexmncn/flemingshop/data/db/ShoppingList.kt
@@ -6,7 +6,7 @@ import java.math.BigInteger
 
 @Entity(tableName = "shoppingList")
 data class ArticleItem(
-    @PrimaryKey(autoGenerate = false) val codebar: Int,
+    @PrimaryKey(autoGenerate = false) val codebar: Long,
     val detalle: String,
     val pvp: Float,
     val quantity: Int
